+++ web/documentserver-example/csharp-mvc/Helpers/Users.cs
@@ -31,7 +31,8 @@ namespace OnlineEditorsExampleMVC.Helpers
             "Can perform all actions with comments",
             "The file favorite state is undefined",
             "Can create files from templates using data from the editor",
-            "Can see the information about all users"
+            "Can see the information about all users",
+            "Has an avatar"
         };
 
         static List<string> descr_user_2 = new List<string>()
@@ -41,7 +42,8 @@ namespace OnlineEditorsExampleMVC.Helpers
             "Can view comments, edit his own comments and comments left by users with no group. Can remove his own comments only",
             "This file is marked as favorite",
             "Can create new files from the editor",
-            "Can see the information about users from Group2 and users who don’t belong to any group"
+            "Can see the information about users from Group2 and users who don’t belong to any group",
+            "Has an avatar"
         };
 
         static List<string> descr_user_3 = new List<string>()
@@ -85,6 +87,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     null,
                     new List<string>(),
                     descr_user_1,
+                    true,
                     true
                 ),
             new User(
@@ -103,7 +106,8 @@ namespace OnlineEditorsExampleMVC.Helpers
                     true,
                     new List<string>(),
                     descr_user_2,
-                    false
+                    false,
+                    true
                 ),
             new User(
                     "uid-3",
@@ -121,6 +125,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     false,
                     new List<string>() { "copy", "download", "print" },
                     descr_user_3,
+                    false,
                     false
                 ),
             new User(
@@ -134,6 +139,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     null,
                     new List<string>() { "protect" },
                     descr_user_0,
+                    false,
                     false
                 )
         };
@@ -171,6 +177,24 @@ namespace OnlineEditorsExampleMVC.Helpers
             }
             return usersData;
         }
+
+        public static List<Dictionary<string, object>> getUsersInfo(string id)
+        {
+            List<Dictionary<string, object>> usersData = new List<Dictionary<string, object>>();
+            if (id != "uid-0") {
+                foreach (User user in users)
+                {
+                    usersData.Add(new Dictionary<string, object>()
+                        {
+                            {"id", user.id},
+                            {"name", user.name },
+                            {"email", user.email },
+                            {"image", user.avatar ? DocManagerHelper.GetServerUrl(false) + "/Content/images/" + user.id + ".png" : null}
+                        });
+                }
+            }
+            return usersData;
+        }
     }
 
     public class User
@@ -186,8 +210,9 @@ namespace OnlineEditorsExampleMVC.Helpers
         public List<string> descriptions;
         public bool templates;
         public List<string> userInfoGroups;
+        public bool avatar;
 
-        public User(string id, string name, string email, string group, List<string> reviewGroups, Dictionary<string, object> commentGroups, List<string> userInfoGroups, bool? favorite, List<string> deniedPermissions, List<string> descriptions, bool templates)
+        public User(string id, string name, string email, string group, List<string> reviewGroups, Dictionary<string, object> commentGroups, List<string> userInfoGroups, bool? favorite, List<string> deniedPermissions, List<string> descriptions, bool templates, bool avatar)
         {
             this.id = id;
             this.name = name;
@@ -200,6 +225,7 @@ namespace OnlineEditorsExampleMVC.Helpers
             this.descriptions = descriptions;
             this.templates = templates;
             this.userInfoGroups = userInfoGroups;
+            this.avatar = avatar;
         }
     }
 }